Struct gltf::accessor::sparse::Sparse

pub struct Sparse<'a> { /* fields omitted */ }

Sparse storage of attributes that deviate from their initialization value.

Methods

impl<'a> Sparse<'a>

pub fn count(&self) -> u32

Returns the number of attributes encoded in this sparse accessor.

pub fn indices(&self) -> Indices<'a>

Returns an index array of size count that points to those accessor attributes that deviate from their initialization value.

pub fn values(&self) -> Values<'a>

Returns an array of size count * number_of_components, storing the displaced accessor attributes pointed by indices.

pub fn extras(&self) -> &Extras

Optional application specific data.