Trait gltf::accessor::util::Item

pub trait Item {
    fn from_slice(slice: &[u8]) -> Self;
    fn zero() -> Self;
}

Represents items that can be read by an Accessor.

Required methods

fn from_slice(slice: &[u8]) -> Self

Create an object of this type from a byte slice.

fn zero() -> Self

Create an object of this type that represents a zero value.