gltf::animation::util
Module morph_target_weights

Casting iterator adapters for morph target weights.

Structs

CastingIter

Casting iterator for MorphTargetWeights.

F32

Type which describes how to cast any weight into f32.

I8

Type which describes how to cast any weight into i8.

I16

Type which describes how to cast any weight into i16.

U8

Type which describes how to cast any weight into u8.

U16

Type which describes how to cast any weight into u16.

Traits

Cast

Trait for types which describe casting behaviour.