Module gltf::animation::util::rotations

Casting iterator adapters for rotations.

Structs

CastingIter

Casting iterator for Rotations.

F32

Type which describes how to cast any weight into f32.

I8

Type which describes how to cast any weight into i8.

I16

Type which describes how to cast any weight into i16.

U8

Type which describes how to cast any weight into u8.

U16

Type which describes how to cast any weight into u16.

Traits

Cast

Trait for types which describe casting behaviour.