gltf
Module material

Material properties of primitives.

Structs

Material

The material appearance of a primitive.

NormalTexture

Defines the normal texture of a material.

OcclusionTexture

Defines the occlusion texture of a material.

PbrMetallicRoughness

A set of parameter values that are used to define the metallic-roughness material model from Physically-Based Rendering (PBR) methodology.

Transmission

A set of parameter values that are used to define the transmissions factor of the material

Volume

Parameter values that define a volume for the transmission of light through the material

Enums

AlphaMode

The alpha rendering mode of a material.