Module gltf::mesh::util::colors

Casting iterator adapters for colors.

Structs

CastingIter
Casting iterator for Colors.

RgbF32
Type which describes how to cast any color into RGB f32.

RgbU8
Type which describes how to cast any color into RGB u8.

RgbU16
Type which describes how to cast any color into RGB u16.

RgbaF32
Type which describes how to cast any color into RGBA f32.

RgbaU8
Type which describes how to cast any color into RGBA u8.

RgbaU16
Type which describes how to cast any color into RGBA u16.

Traits

Cast
Trait for types which describe casting behaviour.