Struct gltf::skin::Skin

pub struct Skin<'a> { /* fields omitted */ }

Joints and matrices defining a skin.

Methods

impl<'a> Skin<'a>

pub fn index(&self) -> usize

Returns the internal JSON index.

pub fn extras(&self) -> &Extras

Optional application specific data.

pub fn inverse_bind_matrices(&self) -> Option<Accessor<'a>>

Returns the accessor containing the 4x4 inverse-bind matrices.

When None, each matrix is assumed to be the 4x4 identity matrix which implies that the inverse-bind matrices were pre-applied.

pub fn reader<'s, F>(&'a self, get_buffer_data: F) -> Reader<'a, 's, F> where
F: Clone + Fn(Buffer<'a>) -> Option<&'s [u8]>,

Constructs a skin reader.

Important traits for Joints<'a>
impl<'a> Iterator for Joints<'a> type Item = Node<'a>;

pub fn joints(&self) -> Joints<'a>

Returns an Iterator that visits the skeleton nodes used as joints in this skin.

pub fn name(&self) -> Option<&str>

Optional user-defined name for this object.

pub fn skeleton(&self) -> Option<Node<'a>>

Returns the node used as the skeleton root. When None, joints transforms resolve to scene root.