Struct gltf::Document

pub struct Document(_);

glTF JSON wrapper.

Methods

impl Document

pub fn from_json(json: Root) -> Result<Self>

Loads glTF from pre-deserialized JSON.

pub fn from_json_without_validation(json: Root) -> Self

Loads glTF from pre-deserialized JSON without performing validation checks.

pub fn into_json(self) -> Root

Unwraps the glTF document.

Important traits for Accessors<'a>
impl<'a> Iterator for Accessors<'a> type Item = Accessor<'a>;

pub fn accessors(&self) -> Accessors

Returns an Iterator that visits the accessors of the glTF asset.

Important traits for Animations<'a>
impl<'a> Iterator for Animations<'a> type Item = Animation<'a>;

pub fn animations(&self) -> Animations

Returns an Iterator that visits the animations of the glTF asset.

Important traits for Buffers<'a>
impl<'a> Iterator for Buffers<'a> type Item = Buffer<'a>;

pub fn buffers(&self) -> Buffers

Returns an Iterator that visits the pre-loaded buffers of the glTF asset.

Important traits for Cameras<'a>
impl<'a> Iterator for Cameras<'a> type Item = Camera<'a>;

pub fn cameras(&self) -> Cameras

Returns an Iterator that visits the cameras of the glTF asset.

pub fn default_scene(&self) -> Option<Scene>

Returns the default scene, if provided.

Important traits for ExtensionsUsed<'a>
impl<'a> Iterator for ExtensionsUsed<'a> type Item = &'a str;

pub fn extensions_used(&self) -> ExtensionsUsed

Returns the extensions referenced in this .document file.

Important traits for ExtensionsRequired<'a>
impl<'a> Iterator for ExtensionsRequired<'a> type Item = &'a str;

pub fn extensions_required(&self) -> ExtensionsRequired

Returns the extensions required to load and render this asset.

Important traits for Images<'a>
impl<'a> Iterator for Images<'a> type Item = Image<'a>;

pub fn images(&self) -> Images

Returns an Iterator that visits the pre-loaded images of the glTF asset.

pub fn lights(&self) -> Option<Lights>

Returns an Iterator that visits the lights of the glTF asset as defined by the KHR_lights_punctual extension.

Important traits for Materials<'a>
impl<'a> Iterator for Materials<'a> type Item = Material<'a>;

pub fn materials(&self) -> Materials

Returns an Iterator that visits the materials of the glTF asset.

Important traits for Meshes<'a>
impl<'a> Iterator for Meshes<'a> type Item = Mesh<'a>;

pub fn meshes(&self) -> Meshes

Returns an Iterator that visits the meshes of the glTF asset.

Important traits for Nodes<'a>
impl<'a> Iterator for Nodes<'a> type Item = Node<'a>;

pub fn nodes(&self) -> Nodes

Returns an Iterator that visits the nodes of the glTF asset.

Important traits for Samplers<'a>
impl<'a> Iterator for Samplers<'a> type Item = Sampler<'a>;

pub fn samplers(&self) -> Samplers

Returns an Iterator that visits the samplers of the glTF asset.

Important traits for Scenes<'a>
impl<'a> Iterator for Scenes<'a> type Item = Scene<'a>;

pub fn scenes(&self) -> Scenes

Returns an Iterator that visits the scenes of the glTF asset.

Important traits for Skins<'a>
impl<'a> Iterator for Skins<'a> type Item = Skin<'a>;

pub fn skins(&self) -> Skins

Returns an Iterator that visits the skins of the glTF asset.

Important traits for Textures<'a>
impl<'a> Iterator for Textures<'a> type Item = Texture<'a>;

pub fn textures(&self) -> Textures

Returns an Iterator that visits the textures of the glTF asset.

Important traits for Views<'a>
impl<'a> Iterator for Views<'a> type Item = View<'a>;

pub fn views(&self) -> Views

Returns an Iterator that visits the pre-loaded buffer views of the glTF asset.