Struct gltf::Node

pub struct Node<'a> { /* fields omitted */ }

A node in the node hierarchy.

When a node contains a skin, all its meshes contain JOINTS_0 and WEIGHTS_0 attributes.

Methods

impl<'a> Node<'a>

pub fn index(&self) -> usize

Returns the internal JSON index.

pub fn camera(&self) -> Option<Camera<'a>>

Returns the camera referenced by this node.

Important traits for Children<'a>
impl<'a> Iterator for Children<'a> type Item = Node<'a>;

pub fn children(&self) -> Children<'a>

Returns an Iterator that visits the node's children.

pub fn extras(&self) -> &'a Extras

Optional application specific data.

pub fn light(&self) -> Option<Light<'a>>

Returns the light at this node as defined by the KHR_lights_punctual extension.

pub fn mesh(&self) -> Option<Mesh<'a>>

Returns the mesh referenced by this node.

pub fn name(&self) -> Option<&'a str>

Optional user-defined name for this object.

pub fn transform(&self) -> Transform

Returns the node's transform.

pub fn skin(&self) -> Option<Skin<'a>>

Returns the skin referenced by this node.

pub fn weights(&self) -> Option<&'a [f32]>

Returns the weights of the instantiated morph target.