gltf
Module texture

Textures and their samplers.

Structs

Info

A reference to a Texture.

Sampler

Texture sampler properties for filtering and wrapping modes.

Texture

A texture and its sampler.

TextureTransform

Many techniques can be used to optimize resource usage for a 3d scene. Chief among them is the ability to minimize the number of textures the GPU must load. To achieve this, many engines encourage packing many objects’ low-resolution textures into a single large texture atlas. The region of the resulting atlas that corresponds with each object is then defined by vertical and horizontal offsets, and the width and height of the region.

Enums

MagFilter

Magnification filter.

MinFilter

Minification filter.

WrappingMode

Texture co-ordinate wrapping mode.