gltf::accessor::util
Trait Item

pub trait Item {
    // Required methods
    fn from_slice(slice: &[u8]) -> Self;
    fn zero() -> Self;
}

Represents items that can be read by an Accessor.

Required Methods

fn from_slice(slice: &[u8]) -> Self

Create an object of this type from a byte slice.

fn zero() -> Self

Create an object of this type that represents a zero value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.