gltf
Module iter

Iterators for walking the glTF node hierarchy.

Structs

Accessors

An Iterator that visits every accessor in a glTF asset.

Animations

An Iterator that visits every animation in a glTF asset.

Buffers

An Iterator that visits every buffer in a glTF asset.

Cameras

An Iterator that visits every camera in a glTF asset.

ExtensionsRequired

An Iterator that visits extension strings required by a glTF asset.

ExtensionsUsed

An Iterator that visits extension strings used by a glTF asset.

Images

An Iterator that visits every pre-loaded image in a glTF asset.

Lights

An Iterator that visits every light in a glTF asset.

Materials

An Iterator that visits every material in a glTF asset.

Meshes

An Iterator that visits every mesh in a glTF asset.

Nodes

An Iterator that visits every node in a glTF asset.

Samplers

An Iterator that visits every sampler in a glTF asset.

Scenes

An Iterator that visits every scene in a glTF asset.

Skins

An Iterator that visits every skin in a glTF asset.

Textures

An Iterator that visits every texture in a glTF asset.

Views

An Iterator that visits every buffer view in a glTF asset.