Struct gltf::Document

pub struct Document(_);

glTF JSON wrapper.

Implementations

impl Document

pub fn from_json(json: Root) -> Result<Self>

Loads glTF from pre-deserialized JSON.

pub fn from_json_without_validation(json: Root) -> Self

Loads glTF from pre-deserialized JSON without performing validation checks.

pub fn into_json(self) -> Root

Unwraps the glTF document.

pub fn accessors(&self) -> Accessors<'_>

Returns an Iterator that visits the accessors of the glTF asset.

pub fn animations(&self) -> Animations<'_>

Returns an Iterator that visits the animations of the glTF asset.

pub fn buffers(&self) -> Buffers<'_>

Returns an Iterator that visits the pre-loaded buffers of the glTF asset.

pub fn cameras(&self) -> Cameras<'_>

Returns an Iterator that visits the cameras of the glTF asset.

pub fn default_scene(&self) -> Option<Scene<'_>>

Returns the default scene, if provided.

pub fn extensions_used(&self) -> ExtensionsUsed<'_>

Returns the extensions referenced in this .document file.

pub fn extensions_required(&self) -> ExtensionsRequired<'_>

Returns the extensions required to load and render this asset.

pub fn images(&self) -> Images<'_>

Returns an Iterator that visits the pre-loaded images of the glTF asset.

pub fn materials(&self) -> Materials<'_>

Returns an Iterator that visits the materials of the glTF asset.

pub fn meshes(&self) -> Meshes<'_>

Returns an Iterator that visits the meshes of the glTF asset.

pub fn nodes(&self) -> Nodes<'_>

Returns an Iterator that visits the nodes of the glTF asset.

pub fn samplers(&self) -> Samplers<'_>

Returns an Iterator that visits the samplers of the glTF asset.

pub fn scenes(&self) -> Scenes<'_>

Returns an Iterator that visits the scenes of the glTF asset.

pub fn skins(&self) -> Skins<'_>

Returns an Iterator that visits the skins of the glTF asset.

pub fn textures(&self) -> Textures<'_>

Returns an Iterator that visits the textures of the glTF asset.

pub fn views(&self) -> Views<'_>

Returns an Iterator that visits the pre-loaded buffer views of the glTF asset.