Struct gltf::Gltf

pub struct Gltf {
    pub document: Document,
    pub blob: Option<Vec<u8>>,
}

glTF JSON wrapper plus binary payload.

Fields
document: Document

The glTF JSON wrapper.

blob: Option<Vec<u8>>

The glTF binary payload in the case of binary glTF.

Implementations

impl Gltf

pub fn open<P>(path: P) -> Result<Self>
where P: AsRef<Path>,

Convenience function that loads glTF from the file system.

pub fn from_reader_without_validation<R>(reader: R) -> Result<Self>
where R: Read + Seek,

Loads glTF from a reader without performing validation checks.

pub fn from_reader<R>(reader: R) -> Result<Self>
where R: Read + Seek,

Loads glTF from a reader.

pub fn from_slice_without_validation(slice: &[u8]) -> Result<Self>

Loads glTF from a slice of bytes without performing validation checks.

pub fn from_slice(slice: &[u8]) -> Result<Self>

Loads glTF from a slice of bytes.